Move shape indicators to overlay util
Replace the React CanvasShapeIndicators component with a new ShapeIndicatorOverlayUtil and wire it into the overlay system. The component and its CSS class (.tl-canvas-indicators) were removed, DefaultCanvas now relies on CanvasOverlays, and the editor package exports the new overlay util. Tests and examples were adjusted to look for .tl-canvas-overlays, and default overlay utils were updated to include ShapeIndicatorOverlayUtil. Also updated a README pointer API reference (currentPagePoint -> getOriginPagePoint).
