+++ packages/editor/editor.css
@@ -959,17 +959,10 @@ input,
 	contain: size layout;
 }
 
-.tl-canvas-indicators {
-	position: absolute;
-	inset: 0;
-	pointer-events: none;
-}
-
 .tl-canvas-overlays {
 	position: absolute;
 	inset: 0;
 	pointer-events: none;
-	z-index: var(--tl-layer-overlays-selection-fg);
 }
 
 /* ---------------------- Shape --------------------- */
